pyproject.toml
[tool.poetry]
name = "narratorx"
version = "0.1.0"
description = ""
authors = ["BedirT <bedirtpkn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
surya-ocr = "^0.6.13"
streamlit = "^1.39.0"
tts = "^0.22.0"
tqdm = "^4.66.6"
pydantic = "^2.9.2"
torch = "^2.5.1"
click = "^8.1.7"
soundfile = "^0.12.1"
pymupdf = "^1.24.13"
litellm = "^1.52.0"
tiktoken = "^0.8.0"
transformers = "^4.46.2"
unstructured = "^0.16.4"
colorlog = "^6.9.0"
ollama = "^0.3.3"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
flake8 = "^7.1.1"
black = "^24.10.0"
pre-commit = "^4.0.1"
isort = "^5.13.2"

[tool.poetry.scripts]
narratorx = "narratorx.cli:main"

[tool.black]
line-length = 100
exclude = '.venv/*'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
